+++ test/integration/nullmap.dart
@@ -21,7 +21,7 @@ void runNullMapTests(String user, String password, String db, int port, String h
     test('read data', () {
       var c = new Completer();
       pool.query('select * from nullmap').then((Results results) {
-        results.stream.listen((row) {
+        results.listen((row) {
           expect(row[0], equals(null));
           expect(row[1].toString(), equals('b'));
           expect(row[2].toString(), equals('c'));