+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+v0.8.0
+* Breaking change: Results no longer has a 'stream' property - it now implements Stream itself.
+As a result, it also no longer has a 'rows' property, or a 'toResultsList()' method - you
+can use 'toList()' to convert it into a list instead.
+
 v0.7.0
 ------
 * Rewritten some connection handling code to make it more robust, and